icon button refactoring
