+++ src/components/IconButton/ViewMonthGraphsButton.tsx
@@ -1,12 +1,13 @@
 import React, { FC } from "react";
 import { faChartSimple } from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
 import { useMonthTotalsContext } from "../../context/MonthTotalsContext";
 import { openModal, setTotalsData } from "../../slices/modalSlice";
 import { useAppDispatch } from "../../store";
 import { IconButtonProps } from "../../types/uiTypes";
 import { ModalCode } from "../../utils/const";
+
+import { GeneralIconButton } from "./GeneralIconButton";
 
 export const ViewMonthGraphsButton: FC<IconButtonProps> = ({ size }) => {
   const dispatch = useAppDispatch();
@@ -24,19 +25,11 @@ export const ViewMonthGraphsButton: FC<IconButtonProps> = ({ size }) => {
   };
 
   return (
-    <span className="hidden sm:block">
-      <button
-        className={"text-black py-2 rounded mr-2 mb-2 sm:mb-0"}
-        onClick={handleShowGraphs}
-        title="month graph"
-      >
-        <div className="sm:hidden">month graph</div>
-        <FontAwesomeIcon
-          icon={faChartSimple}
-          className="hidden sm:block"
-          size={size}
-        />
-      </button>
-    </span>
+    <GeneralIconButton
+      onClick={handleShowGraphs}
+      title="month graph"
+      icon={faChartSimple}
+      size={size}
+    />
   );
 };
