+++ src/components/IconButton/ViewPieChartButton.tsx
@@ -1,12 +1,13 @@
 import React, { FC } from "react";
 import { faPieChart } from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
 import { useMonthTotalsContext } from "../../context/MonthTotalsContext";
 import { openModal, setTotalsData } from "../../slices/modalSlice";
 import { useAppDispatch } from "../../store";
 import { IconButtonProps } from "../../types/uiTypes";
 import { ModalCode } from "../../utils/const";
+
+import { GeneralIconButton } from "./GeneralIconButton";
 
 export const ViewPieChartButton: FC<IconButtonProps> = ({ size }) => {
   const dispatch = useAppDispatch();
@@ -24,19 +25,11 @@ export const ViewPieChartButton: FC<IconButtonProps> = ({ size }) => {
   };
 
   return (
-    <span className="hidden sm:block">
-      <button
-        className={"text-black py-2 rounded mr-2 mb-2 sm:mb-0"}
-        onClick={handleShowPieChart}
-        title="pie charts"
-      >
-        <div className="sm:hidden">pie charts</div>
-        <FontAwesomeIcon
-          icon={faPieChart}
-          className="hidden sm:block"
-          size={size}
-        />
-      </button>
-    </span>
+    <GeneralIconButton
+      onClick={handleShowPieChart}
+      title="pie charts"
+      icon={faPieChart}
+      size={size}
+    />
   );
 };
