+++ src/components/Transactions/TransactionMonthFooter.tsx
@@ -1,8 +1,8 @@
 import React, { FC } from "react";
 
-import { ViewMonthGraphsButton } from "../Analytics/ViewMonthGraphsButton";
-import { ViewPieChartButton } from "../Analytics/ViewPieChartButton";
-import { ViewTotalsButton } from "../Analytics/ViewTotalsButton";
+import { ViewMonthGraphsButton } from "../IconButton/ViewMonthGraphsButton";
+import { ViewPieChartButton } from "../IconButton/ViewPieChartButton";
+import { ViewTotalsButton } from "../IconButton/ViewTotalsButton";
 import { ViewAmount } from "../Util/ViewAmount";
 
 interface TransactionMonthFooterProps {
